requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.3
idna==2.9
lxml==4.5.0
multidict==4.7.5
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
tzlocal==2.0.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
fake_useragent==0.1.11